+++ src/kernel/multitasking/sigset.h
@@ -5,4 +5,4 @@
 sigset_t sigset_bitwise_or(sigset_t s1, sigset_t s2);
 sigset_t sigset_bitwise_and(sigset_t s1, sigset_t s2);
 sigset_t sigset_bitwise_not(sigset_t s);
-bool sigset_is_bit_set(sigset_t s, int bit);
+bool sigset_is_bit_set(sigset_t s, size_t bit);
